# Conflicts:
#	src/main/kotlin/theater/controller/EmployeeController.kt
#	src/main/kotlin/theater/dao/ActorsDao.kt
#	src/main/kotlin/theater/service/EmployeeService.kt

+++ src/main/kotlin/theater/model/Employee.kt
@@ -2,6 +2,6 @@ package theater.model
 
 import java.sql.Date
 
-data class Employee(var id: Long, val fio: String, val sex: String,
+data class Employee(var id: Int, val fio: String, val sex: String,
                     val birthDate: Date, val childrenAmount: Int, val salary: Int,
                     val origin: String, val hireDate: Date)
